Move Product.{Contact,TermsOfUse} inner classes to src/groovy
