+++ src/groovy/com/unifina/api/CreateProductCommand.groovy
@@ -2,7 +2,9 @@ package com.unifina.api
 
 import com.unifina.domain.data.Stream
 import com.unifina.domain.marketplace.Category
+import com.unifina.domain.marketplace.Contact
 import com.unifina.domain.marketplace.Product
+import com.unifina.domain.marketplace.TermsOfUse
 import grails.validation.Validateable
 
 @Validateable
@@ -23,8 +25,8 @@ class CreateProductCommand {
 	Long pricePerSecond = 0
 	Product.Currency priceCurrency = Product.Currency.DATA
 	Long minimumSubscriptionInSeconds = 0
-	Product.Contact contact = new Product.Contact()
-	Product.TermsOfUse termsOfUse = new Product.TermsOfUse()
+	Contact contact = new Contact()
+	TermsOfUse termsOfUse = new TermsOfUse()
 
 	static constraints = {
 		importFrom(Product)
