+++ src/groovy/com/unifina/api/UpdateProductCommand.groovy
@@ -2,7 +2,9 @@ package com.unifina.api
 
 import com.unifina.domain.data.Stream
 import com.unifina.domain.marketplace.Category
+import com.unifina.domain.marketplace.Contact
 import com.unifina.domain.marketplace.Product
+import com.unifina.domain.marketplace.TermsOfUse
 import com.unifina.domain.security.SecUser
 import com.unifina.service.PermissionService
 import grails.compiler.GrailsCompileStatic
@@ -20,8 +22,8 @@ class UpdateProductCommand {
 	Stream previewStream
 	String previewConfigJson
 	Map<String, Object> pendingChanges
-	Product.Contact contact = new Product.Contact()
-	Product.TermsOfUse termsOfUse = new Product.TermsOfUse()
+	Contact contact = new Contact()
+	TermsOfUse termsOfUse = new TermsOfUse()
 
 	// Below are used only when updating NOT_DEPLOYED product
 	String ownerAddress
