+++ test/integration/com/unifina/service/ProductServiceIntegrationSpec.groovy
@@ -66,8 +66,6 @@ class ProductServiceIntegrationSpec extends IntegrationSpec {
 			blockNumber: 40000,
 			blockIndex: 30,
 			owner: troll,
-			contact: new Product.Contact(),
-			termsOfUse: new Product.TermsOfUse(),
 		)
 		p1.save(failOnError: true, validate: true)
 		p2 = new Product(
@@ -82,8 +80,6 @@ class ProductServiceIntegrationSpec extends IntegrationSpec {
 			blockNumber: 40000,
 			blockIndex: 30,
 			owner: troll,
-			contact: new Product.Contact(),
-			termsOfUse: new Product.TermsOfUse(),
 		)
 		p2.save(failOnError: true, validate: true)
